feat: 流水线级回调与流水线编排解耦 #11283

+++ src/backend/ci/core/process/biz-base/src/main/kotlin/com/tencent/devops/process/engine/service/PipelineRepositoryService.kt
@@ -2216,13 +2216,15 @@ class PipelineRepositoryService constructor(
             projectId = projectId,
             pipelineId = pipelineId
         ).map { it.name }.toSet()
-        val needDelNames = existEventNames.subtract(events.keys).toSet()
-        pipelineCallbackDao.delete(
-            dslContext = dslContext,
-            projectId = projectId,
-            pipelineId = pipelineId,
-            names = needDelNames
-        )
+        if (existEventNames.isNotEmpty()) {
+            val needDelNames = existEventNames.subtract(events.keys).toSet()
+            pipelineCallbackDao.delete(
+                dslContext = dslContext,
+                projectId = projectId,
+                pipelineId = pipelineId,
+                names = needDelNames
+            )
+        }
         // 保存回调事件
         pipelineCallbackDao.save(
             dslContext = dslContext,
